requirements.txt
openai>=1.0.0
colorama>=0.4.6
pwinput>=1.0.3
python-dotenv>=1.0.0
rich>=13.3.5
gradio>=4.37.2
fastapi>=0.110.0
uvicorn>=0.29.0
requests>=2.31.0